Split methods and polls
Fix #5
